requirements.txt
streamlit>=1.22.0
langchain>=1.0.0
langchain-community>=0.4.0
langchain-core>=1.0.0
langchain-openai>=1.0.0
langchain-text-splitters>=1.0.0
langchain-classic>=1.0.0
openai>=2.0.0
faiss-cpu>=1.7.4
pdfplumber>=0.7.5
python-docx>=0.8.11
sentence-transformers>=2.2.2
tqdm>=4.64.0
python-multipart>=0.0.5
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0